script/ folder is not included in package -> move to data/
