+++ examples/bosons-fermions-pimd/data/__init__.py
@@ -1,0 +1,8 @@
+"""Support code for the bosons/fermions PIMD tutorial.
+
+* :mod:`~data.analysis` -- reading i-PI output, exact reference energies, and
+  the sign-weighted fermionic estimator (the physics/analysis of the tutorial);
+* :mod:`~data.ipi_runs` -- launching i-PI and its force driver (plumbing kept
+  out of the notebook);
+* :mod:`~data.plots` -- drawing the tutorial figures.
+"""
