+++ examples/bosons-fermions-pimd/data/analysis.py
