+++ examples/bosons-fermions-pimd/data/plots.py
