--- examples/bosons-fermions-pimd/scripts/__init__.py
@@ -1,8 +1,0 @@
-"""Support code for the bosons/fermions PIMD tutorial.
-
-* :mod:`~scripts.analysis` -- reading i-PI output, exact reference energies, and
-  the sign-weighted fermionic estimator (the physics/analysis of the tutorial);
-* :mod:`~scripts.ipi_runs` -- launching i-PI and its force driver (plumbing kept
-  out of the notebook);
-* :mod:`~scripts.plots` -- drawing the tutorial figures.
-"""
